+++ packages/plugin-application-builder/src/fixtures/sample-context/src/create-user.command.ts
@@ -5,6 +5,6 @@ export class CreateUserCommand extends Command<
     { role: string }
 > {
     constructor(payload: { name: string }, sc: { role: string }) {
-        super(payload, sc);
+        super(payload, { securityContext: sc });
     }
 }